+++ src/cmd_pgr/plot/mod.rs
@@ -1,8 +1,10 @@
 use clap::{ArgMatches, Command};
 
 pub mod dot;
+pub mod heat;
 pub mod hh;
 pub mod nrps;
+pub mod spectra;
 pub mod venn;
 /// Build the clap subcommand for plot.
 pub fn make_subcommand() -> Command {
@@ -10,16 +12,20 @@ pub fn make_subcommand() -> Command {
         .about("Plots figures")
         .subcommand_required(true)
         .subcommand(hh::make_subcommand())
+        .subcommand(heat::make_subcommand())
         .subcommand(dot::make_subcommand())
         .subcommand(nrps::make_subcommand())
+        .subcommand(spectra::make_subcommand())
         .subcommand(venn::make_subcommand())
 }
 /// Execute the plot command.
 pub fn execute(args: &ArgMatches) -> anyhow::Result<()> {
     match args.subcommand() {
         Some(("hh", sub_matches)) => hh::execute(sub_matches),
+        Some(("heat", sub_matches)) => heat::execute(sub_matches),
         Some(("dot", sub_matches)) => dot::execute(sub_matches),
         Some(("nrps", sub_matches)) => nrps::execute(sub_matches),
+        Some(("spectra", sub_matches)) => spectra::execute(sub_matches),
         Some(("venn", sub_matches)) => venn::execute(sub_matches),
         _ => Ok(()),
     }