+++ src/libs/plot/mod.rs
@@ -2,6 +2,8 @@
 
 pub mod common;
 pub mod dot;
+pub mod heat;
 pub mod histogram;
 pub mod nrps;
+pub mod spectra;
 pub mod venn;